feat: split crossRefValidator into per-slice modules (Lift 15 / #255)
Pre-Lift 15, `path-analyser/src/ontology/crossRefValidator.ts` (675 lines)
hand-encoded the merged-domain shape as zod `.passthrough()` schemas — a
parallel encoding of the per-slice TBoxes (`runtimeStatesSchema.ts`,
`semanticsSchema.ts`, `artifactKindsSchema.ts`,
`globalContextSeedsSchema.ts`). The passthrough schemas did not actually
validate — they only narrowed types — and silently drifted whenever a
slice TBox added a property. No build-time guard caught it.

Adopts option (b) from the issue's acceptance criterion: per-slice
cross-reference modules composed by the validator, with a build-time
check that every slice has a corresponding module.

- New `path-analyser/src/ontology/crossRef/` directory with one module
  per `*Schema.ts` slice:
  - `edgeCrossRef.ts` — no checks (resolves against operation graph,
    not DomainSemantics)
  - `entityKindsCrossRef.ts` — no checks (resolves against operation
    graph, not DomainSemantics)
  - `artifactKindsCrossRef.ts` — checkArtifactKindStateDeclared,
    checkArtifactKindWitnessDeclared
  - `runtimeStatesCrossRef.ts` — checkSemanticBindingTargetResolves,
    checkDisjunctionNotWitnessRedundant,
    checkDisjunctionMemberResolves, checkEventualStateWitnessShape
  - `semanticsCrossRef.ts` — checkSemanticTypeWitnessTargetResolves,
    checkAttributeKindClientMintedPairing
  - `globalContextSeedsCrossRef.ts` — checkGlobalContextSeedsCoherent
- `crossRefValidator.ts` slimmed from 675 → ~290 lines: now a pure
  composer over `CROSS_REF_MODULES`. The hand-encoded structural zod
  shapes are gone — `DomainSemantics` from `types.ts` is the typed view;
  structural validation already happens in `loader.ts` (ajv) against
  the source-of-truth `*Schema.ts` TBoxes. `GlobalContextSeedSchema` is
  kept (it backs the public-boundary `assertSafeGlobalContextSeeds`).
- Behaviour-preserving: every cross-ref check's `code` and `message`
  text is verbatim from the pre-refactor version. Pipeline output is
  unchanged (186 feature.specs + 67 variant.specs). All 517 prior tests
  still pass.

Build-time guard (`configs/camunda-oca/regression-invariants.test.ts`,
new "cross-ref module coverage (Lift 15 / #255)" describe block, three
invariants):

1. "every per-slice TBox under path-analyser/src/ontology/*Schema.ts
   has a corresponding cross-ref module registered in CROSS_REF_MODULES"
   — enumerates `*Schema.ts` filenames and asserts each slice stem
   appears in CROSS_REF_MODULES. Fails with a named-slice message
   ("Adding a TBox slice without considering cross-reference invariants
   is the drift defect Lift 15 / #255 was filed to prevent...") when a
   slice ships without a module.
2. "every cross-ref module with no checks ships an explicit
   noChecksRationale" — prevents an empty module from being
   indistinguishable from a forgotten/unfinished cross-ref.
3. "every registered cross-ref module corresponds to an actual
   *Schema.ts slice file" — catches the inverse (renamed/deleted
   slice).

Red-proof verified locally: commenting out RUNTIME_STATES_CROSS_REF from
the registry fires invariant (1) with the named-slice message.
